OSX fixes
Summary:
Fixes to make OS X build compile.

* The MessageType ostream formatter was generalized for all enums and moved to testharness because we don't want i/ostream stuff in the client binary.
* Some formatting fixes, probably caused by making MessageType an enum class.

Test Plan: make check

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D38553

+++ src/copilot/worker.cc
@@ -118,7 +118,7 @@ void CopilotWorker::CommandCallback(CopilotWorkerCommand command) {
     default: {
         LOG_WARN(options_.info_log,
             "Unexpected message type in copilot worker %d",
-            msg->GetMessageType());
+            static_cast<int>(msg->GetMessageType()));
       }
       break;
     }
